refacto with FacetButton
